backend/requirements.txt
# ClearSky Text-to-SQL Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# AWS SDK
boto3==1.34.25
botocore==1.34.25

# LlamaIndex for LLM orchestration
llama-index==0.10.12
llama-index-llms-bedrock==0.1.3
llama-index-embeddings-bedrock==0.1.3

# Database connectivity (for pgvector - optional)
psycopg[binary]==3.1.17
pgvector==0.2.4

# Local vector store (FAISS)
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
numpy>=1.24.0

# Logging and monitoring
structlog==24.1.0
